requirements.txt
Authlib==1.2.1
blinker==1.6.3
cachelib==0.10.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
connectorx==0.3.2
cryptography==41.0.4
Flask==3.0.0
Flask-OAuthlib==0.9.6
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.26.1
oauthlib==2.1.0
pandas==2.1.1
pyaml==23.9.7
pycparser==2.21
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.1.0
six==1.16.0
SQLAlchemy==2.0.22
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
Werkzeug==2.2.2
zipp==3.17.0
